Added Running bash script as user
Added the ability to run Bash scripts inside the newly created chroot as user instead of root.

Added a new script to create the user without root permission inside the chroot

The ending of lines is now in LF instead of CRLF on the Github (see .gitattributes)

+++ ChrootBuilder/ChrootBuilder.bash
@@ -34,30 +34,28 @@ chrootScriptsDir=ChrootScripts
 
 
 #### Création du groupe chrooted et ajout de celui-ci dans les configurations de SSH ####
-if [ -z "$(cat /etc/ssh/sshd_config | grep "Match Group ${chrootGroup}")" ]
-	then
-		echo "${chrootGroup} did not exist. [...]"
-		# Create chroot group
-		addgroup $chrootGroup
-		
-		# Add chroot group to sshd_config
-		echo -e '\n' | sudo tee -a /etc/ssh/sshd_config > /dev/null
-		echo -e "Match Group ${chrootGroup}" | sudo tee -a /etc/ssh/sshd_config > /dev/null
-		echo -e "\tChrootDirectory ${chrootDir}/%u" | sudo tee -a /etc/ssh/sshd_config > /dev/null
-		
-		echo "${chrootGroup} added. [OK]"
-		echo "${chrootGroup} added to sshd_config. [OK]"
+if [ -z "$(cat /etc/ssh/sshd_config | grep "Match Group ${chrootGroup}")" ]; then
+	echo "${chrootGroup} did not exist. [...]"
+	# Create chroot group
+	addgroup $chrootGroup
+	
+	# Add chroot group to sshd_config
+	echo -e '\n' | sudo tee -a /etc/ssh/sshd_config > /dev/null
+	echo -e "Match Group ${chrootGroup}" | sudo tee -a /etc/ssh/sshd_config > /dev/null
+	echo -e "\tChrootDirectory ${chrootDir}/%u" | sudo tee -a /etc/ssh/sshd_config > /dev/null
+	
+	echo "${chrootGroup} added. [OK]"
+	echo "${chrootGroup} added to sshd_config. [OK]"
 fi
 
 
 
 #### Création du répertoire de stockage des chroot ####
-if [ -d $chrootDir ]
-	then
-		echo "${chrootDir} exist. [OK]"
-	else
-		echo "${chrootDir} did not exist. mkdir ${chrootDir}"
-		mkdir $chrootDir
+if [ -d $chrootDir ]; then
+	echo "${chrootDir} exist. [OK]"
+else
+	echo "${chrootDir} did not exist. mkdir ${chrootDir}"
+	mkdir $chrootDir
 fi
 
 chrootID=$(expr $(ls $chrootDir | wc -l) + 1) # Identifiant de la chroot
@@ -73,16 +71,15 @@ echo -e '\n'
 userDir=$chrootDir/$username 			# Répertoire du chroot de l'utilisateur
 
 # Création de l'utilisateur, ajout dans le goupe chroot et changement de son shell pour Bash
-if [ -n "$(cat /etc/passwd | grep "${username}:")" ]
-	then
-		echo "${username} exist. [...]"
-		usermod -a -G $chrootGroup -s $chrootShell $username # Create user with chrooted group and chroot shell
-		echo "${username} added to ${chrootGroup}. [OK]"
-		echo "${username} shell change to ${chrootShell}. [OK]"
-	else
-		echo "${username} did not exist. [...]"
-		useradd -G $chrootGroup -s $chrootShell $username # Add user to chrooted and change shell to chroot shell
-		echo "${username} added. [OK]"
+if [ -n "$(cat /etc/passwd | grep "${username}:")" ]; then
+	echo "${username} exist. [...]"
+	usermod -a -G $chrootGroup -s $chrootShell $username # Create user with chrooted group and chroot shell
+	echo "${username} added to ${chrootGroup}. [OK]"
+	echo "${username} shell change to ${chrootShell}. [OK]"
+else
+	echo "${username} did not exist. [...]"
+	useradd -G $chrootGroup -s $chrootShell $username # Add user to chrooted and change shell to chroot shell
+	echo "${username} added. [OK]"
 fi
 
 # Change the password of the user
@@ -91,39 +88,35 @@ echo -e $username:$password | chpasswd
 
 
 #### Supprimer les anciennes installation du chroot de l'utilisateur ####
-if [ -d $userDir ]
-	then
-		echo "${userDir} exist. Remove it (Can take a while) [...]"
+if [ -d $userDir ]; then
+	echo "${userDir} exist. Remove it (Can take a while) [...]"
 
-		# Unmount every mounted directories inside the chroot
-		for mtabContent in $(cat /etc/mtab  | grep $userDir)
-		do
-			mountedDir="$(echo $mtabContent | grep $userDir)"
-			if [ -n "${mountedDir}" ] # File/Directory exist
-			then
-				echo "${mountedDir} mounted. Unmount it. [...]"
-				#fuser -skm $mountedDir # Kill every process using it (also kill this process)
-				umount $mountedDir;
-				
-				if mountpoint -q $mountedDir
-					then
-						echo "Cant unmount ${mountedDir}. [ERROR]"
-						echo "Use <fuser -cuk ${mountedDir}> to kill the process using it."
-						echo "or/and <fuser -skm ${mountedDir}> to kill the process using it."
-						echo "or/and <lsof | grep ${mountedDir}> to list the process and manually kill them"
-						exit 1
-				fi
-				
-				echo "${mountedDir} unmounted. [OK]"
+	# Unmount every mounted directories inside the chroot
+	for mtabContent in $(cat /etc/mtab  | grep $userDir); do
+		mountedDir="$(echo $mtabContent | grep $userDir)"
+		if [ -n "${mountedDir}" ]; then # File/Directory exist
+			echo "${mountedDir} mounted. Unmount it. [...]"
+			#fuser -skm $mountedDir # Kill every process using it (also kill this process)
+			umount $mountedDir;
+			
+			if mountpoint -q $mountedDir; then
+				echo "Cant unmount ${mountedDir}. [ERROR]"
+				echo "Use <fuser -cuk ${mountedDir}> to kill the process using it."
+				echo "or/and <fuser -skm ${mountedDir}> to kill the process using it."
+				echo "or/and <lsof | grep ${mountedDir}> to list the process and manually kill them"
+				exit 1
 			fi
-		done
+			
+			echo "${mountedDir} unmounted. [OK]"
+		fi
+	done
 
-		# Remove user chroot
-		rm -r $userDir
-		
-		echo "${userDir} removed. [OK]"
-	else
-		echo "${userDir} did not exist. [OK]"
+	# Remove user chroot
+	rm -r $userDir
+	
+	echo "${userDir} removed. [OK]"
+else
+	echo "${userDir} did not exist. [OK]"
 fi
 
 
@@ -141,18 +134,16 @@ echo -e "type=directory" | tee -a /etc/schroot/chroot.d/$username.conf > /dev/nu
 #### Installer Ubuntu dans le chroot (Copie du Template) ####
 # Remove process management from Template (Need to be mounted later, not copied)
 # Unmount
-if mountpoint -q $chrootTemplateDir/proc
-	then
-		echo "${chrootTemplateDir}/proc mounted. [...]"
-		umount $chrootTemplateDir/proc
+if mountpoint -q $chrootTemplateDir/proc; then
+	echo "${chrootTemplateDir}/proc mounted. [...]"
+	umount $chrootTemplateDir/proc
 fi
 echo "${chrootTemplateDir}/proc unmounted. [OK]"
 
 # Remove
-if [ -d $chrootTemplateDir/proc ]
-	then
-		echo "${chrootTemplateDir}/proc exist. [...]"
-		rm -r $chrootTemplateDir/proc
+if [ -d $chrootTemplateDir/proc ]; then
+	echo "${chrootTemplateDir}/proc exist. [...]"
+	rm -r $chrootTemplateDir/proc
 fi
 echo "${chrootTemplateDir}/proc removed. [OK]"
 
@@ -167,6 +158,11 @@ cp --parents /run/systemd/resolve/stub-resolv.conf $userDir
 cp --parents /etc/resolv.conf $userDir
 echo -e "$(hostname -I)\t$(hostname)" >> $userDir/etc/hosts # Ajouter le hostname et l'ip attitré dans le chroot
 
+# Configure permission of the chroot
+chmod -R 777 $userDir
+chown root:root $userDir
+chmod 0755 $userDir
+
 # Add process management
 mkdir $userDir/proc
 mount -o bind /proc $userDir/proc
@@ -178,8 +174,26 @@ mount -o bind /dev/pts $userDir/dev/pts
 for scriptPath in $chrootScriptsDir/*.bash; do
     [ -f "$scriptPath" ] || break	# Break if the file didnt exist
 	scriptName=${scriptPath#"$chrootScriptsDir/"} # Remove $chrootScriptsDir/ to get the name
-    cp -v $scriptPath $userDir
-	chroot $userDir $chrootShell $scriptName $username $password $(id -u $username) $(id -g $username) $chrootID
+	answer=0
+	
+	while [ $answer != "Y" ] && [ $answer != "N" ]; do
+		read -p "Exécuter ${scriptName} [Y]es/[N]o : " answer # Exécution du script?
+	done
+	
+	if [ $answer == "Y" ]; then
+		answer=0
+		while [ $answer != "Y" ] && [ $answer != "N" ]; do
+			read -p "Exécuter ${scriptName} en Root [Y]es/[N]o : " answer # Exécution du script en Root?
+		done
+		
+		if [ $answer == "Y" ]; then
+			cp -v $scriptPath $userDir
+			chroot --userspec=root:root $userDir $chrootShell $scriptName $username $password $(id -u $username) $(id -g $username) $chrootID
+		else
+			cp -v $scriptPath $userDir
+			chroot --userspec=$username:$username $userDir $chrootShell $scriptName $username $password $(id -u $username) $(id -g $username) $chrootID
+		fi
+	fi
 done
